requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.42.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-mongodb>=0.1.0
motor>=3.3.1
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
plaid-python>=16.0.0
python-multipart>=0.0.6
pillow>=10.0.0
pytesseract>=0.3.10
httpx>=0.24.0
